+++ src/__tests__/changemakers.int.test.ts
@@ -1126,6 +1126,12 @@ describe('/changemakers', () => {
 			const grants = await loadPermissionGrantBundle(
 				db,
 				getAuthContext(systemUser, true),
+				undefined,
+				undefined,
+				undefined,
+				undefined,
+				undefined,
+				undefined,
 				NO_LIMIT,
 				NO_OFFSET,
 			);