+++ src/database/operations/permissionGrants/loadPermissionGrantBundle.ts
@@ -1,9 +1,29 @@
 import { generateLoadBundleOperation } from '../generators';
-import type { PermissionGrant } from '../../../types';
+import type {
+	Id,
+	PermissionGrant,
+	PermissionGrantGranteeType,
+	PermissionGrantVerb,
+	ShortCode,
+} from '../../../types';
 
 const loadPermissionGrantBundle = generateLoadBundleOperation<
 	PermissionGrant,
-	[]
->('permissionGrants.selectWithPagination', []);
+	[
+		changemakerId: Id | undefined,
+		funderShortCode: ShortCode | undefined,
+		dataProviderShortCode: ShortCode | undefined,
+		proposalId: Id | undefined,
+		granteeType: PermissionGrantGranteeType | undefined,
+		verb: PermissionGrantVerb | undefined,
+	]
+>('permissionGrants.selectWithPagination', [
+	'changemakerId',
+	'funderShortCode',
+	'dataProviderShortCode',
+	'proposalId',
+	'granteeType',
+	'verb',
+]);
 
 export { loadPermissionGrantBundle };